requirements.txt
flask
ultralytics
torch
torchvision
pillow
numpy
opencv-python
